Music base
I started working on setting up a music base, but ytdl got taken down while I was working on this. I haven't really felt motivated to go back as a result. I also found out that the audio wouldn't work if I hosted the bot through certain websites with the free tiers. So eh. Not as motivated to finish it.